Refactoring Build System (#223)
* using Petabridge.Templates build system

* close #199 - referenced Akka.NET NuGet package directly in dependencies test

* close #224 - use dotnet run for post-build integration testing of NBench.Runner

+++ tests/NBench.Tests.Reporting/Reporting/MarkdownBenchmarkOutputSpec.cs
@@ -54,7 +54,7 @@ namespace NBench.Tests.End2End.Reporting
             var fakeBenchmarkResults = new BenchmarkResults("NBench.FakeBenchmark",
                 new BenchmarkSettings(TestMode.Test, RunMode.Iterations, 30, 1000,
                     new List<IBenchmarkSetting>(),
-                    new ConcurrentDictionary<MetricName, MetricsCollectorSelector>()),
+                    new Dictionary<MetricName, MetricsCollectorSelector>()),
                 new List<BenchmarkRunReport>()
                 {
                     new BenchmarkRunReport(TimeSpan.FromSeconds(3),
@@ -82,10 +82,13 @@ namespace NBench.Tests.End2End.Reporting
 
         private static void CleanPerfDir(string path)
         {
-            var files = Directory.GetFiles(path, "NBench.FakeBenchmark*", SearchOption.AllDirectories);
-            foreach (var file in files)
+            if (Directory.Exists(path)) // no need to clean the directory if we don't have it
             {
-                File.Delete(file);
+                var files = Directory.GetFiles(path, "NBench.FakeBenchmark*", SearchOption.AllDirectories);
+                foreach (var file in files)
+                {
+                    File.Delete(file);
+                }
             }
         }
